+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Office 365/Bases/PaletteOffice365BlueThemeDarkModeBase.cs
@@ -146,12 +146,12 @@ namespace Krypton.Toolkit
                                                                          };
         private static readonly Color[] _buttonBackColors = { Color.FromArgb(250, 250, 250), // Button, Disabled, Back 1
                                                                          Color.FromArgb(250, 250, 250), // Button, Disabled, Back 2
-                                                                         Color.FromArgb(248, 225, 135), // Button, Tracking, Back 1
-                                                                         Color.FromArgb(251, 248, 224), // Button, Tracking, Back 2
-                                                                         Color.FromArgb(255, 228, 138), // Button, Pressed, Back 1
-                                                                         Color.FromArgb(194, 118, 43),  // Button, Pressed, Back 2
-                                                                         Color.FromArgb(255, 216, 108), // Button, Checked, Back 1
-                                                                         Color.FromArgb(255, 244, 128), // Button, Checked, Back 2
+                                                                         Color.FromArgb(187, 206, 230), // Button, Tracking, Back 1
+                                                                         Color.FromArgb(174, 194, 219), // Button, Tracking, Back 2
+                                                                         Color.FromArgb(187, 206, 230), // Button, Pressed, Back 1
+                                                                         Color.FromArgb(174, 194, 219),  // Button, Pressed, Back 2
+                                                                         Color.FromArgb(187, 206, 230), // Button, Checked, Back 1
+                                                                         Color.FromArgb(174, 194, 219), // Button, Checked, Back 2
                                                                          Color.FromArgb(255, 225, 104), // Button, Checked Tracking, Back 1
                                                                          Color.FromArgb(255, 249, 196)  // Button, Checked Tracking, Back 2
                                                                        };